code: move dashboard report edit modal to new component
