+++ src/components/dashboard-edit/styles.module.scss
@@ -26,9 +26,7 @@
   text-decoration: none;
 }
 
-.dashboardEditModuleWrapper {
-  margin-top: 24px;
-}
+.dashboardEditModuleWrapper { margin-top: 24px; }
 
 .arrowButtons {
   margin-right: 8px;
@@ -37,130 +35,7 @@
 }
 .arrowButtons:hover { opacity: 1; }
 
-.name {
-  margin-left: 8px;
-}
-
-.reportModalReportList {
-  overflow-y: auto;
-  margin-top: 8px;
-  padding-left: 24px;
-  padding-right: 24px;
-  flex-grow: 1;
-  flex-shrink: 1;
-}
-
-.parent {
-  display: flex;
-  flex-grow: 1;
-  flex-shrink: 1;
-}
-.left {
-  display: flex;
-  flex-direction: column;
-  flex-shrink: 1;
-  flex-grow: 1;
-  width: 390px;
-  height: 100%;
-}
-.right {
-  display: flex;
-  flex-direction: column;
-  width: 600px;
-  height: 100%;
-  border-left: 1px solid $grayLight;
-}
-
-.searchBar {
-  background-color: $grayLightest;
-  flex-grow: 0;
-  flex-shrink: 0;
-}
-
-.reportTypeList {
-  flex-grow: 1;
-  flex-shrink: 1;
-
-  overflow-y: auto;
-  -webkit-overflow-scrolling: touch;
-}
-
-.reportTypeItem {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  cursor: pointer;
-
-  padding-top: 12px;
-  padding-left: 24px;
-  padding-right: 24px;
-
-  border-bottom: 1px solid $grayLight;
-  user-select: none;
-  outline: none;
-  height: 120px;
-}
-.reportTypeItem:focus {
-  background-color: $grayLightest;
-}
-
-.nameRow {
-  display: flex;
-  align-items: center;
-}
-
-.name {
-  margin-left: 16px;
-  font-weight: 500;
-  height: 16px;
-}
-
-.reportTypeItemDescription {
-  margin-top: 8px;
-  margin-right: 24px;
-}
-
-.reportTypeImage {
-  height: 110px;
-}
-
-.reportImageBackdrop {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 200px;
-  width: 100%;
-  background-color: $grayLightest;
-}
-
-.reportFullDescription {
-  padding: 24px;
-  white-space: pre-wrap;
-}
-
-.reportConfigurationParent {
-  flex-grow: 1;
-  flex-shrink: 1;
-  padding-left: 24px;
-  padding-right: 24px;
-  padding-top: 12px;
-  overflow-y: auto;
-  -webkit-overflow-scrolling: touch;
-}
-
-.reportBackdrop {
-  flex-grow: 1;
-  flex-shrink: 1;
-  background-color: $grayLightest;
-  padding: 24px;
-  overflow-y: auto;
-}
-
-.reportModalWrapper {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
+.name { margin-left: 8px; }
 
 .dangerZoneWrapper { display: flex; }
 .dangerZoneLeft {
