+++ src/components/dashboard-report-edit-modal/README.md
@@ -1,0 +1,3 @@
+# DashboardReportEditModal
+
+A modal used to edit reports within the dashboard edit page.
