+++ src/components/dashboard-report-edit-modal/styles.module.scss
@@ -1,0 +1,123 @@
+@import "@density/ui/variables/colors.scss";
+@import "@density/ui/variables/spacing.scss";
+
+.reportModalReportList {
+  overflow-y: auto;
+  margin-top: 8px;
+  padding-left: 24px;
+  padding-right: 24px;
+  flex-grow: 1;
+  flex-shrink: 1;
+}
+
+.parent {
+  display: flex;
+  flex-grow: 1;
+  flex-shrink: 1;
+}
+.left {
+  display: flex;
+  flex-direction: column;
+  flex-shrink: 1;
+  flex-grow: 1;
+  width: 390px;
+  height: 100%;
+}
+.right {
+  display: flex;
+  flex-direction: column;
+  width: 600px;
+  height: 100%;
+  border-left: 1px solid $grayLight;
+}
+
+.searchBar {
+  background-color: $grayLightest;
+  flex-grow: 0;
+  flex-shrink: 0;
+}
+
+.reportTypeList {
+  flex-grow: 1;
+  flex-shrink: 1;
+
+  overflow-y: auto;
+  -webkit-overflow-scrolling: touch;
+}
+
+.reportTypeItem {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  cursor: pointer;
+
+  padding-top: 12px;
+  padding-left: 24px;
+  padding-right: 24px;
+
+  border-bottom: 1px solid $grayLight;
+  user-select: none;
+  outline: none;
+  height: 120px;
+}
+.reportTypeItem:focus {
+  background-color: $grayLightest;
+}
+
+.nameRow {
+  display: flex;
+  align-items: center;
+}
+
+.name {
+  margin-left: 16px;
+  font-weight: 500;
+  height: 16px;
+}
+
+.reportTypeItemDescription {
+  margin-top: 8px;
+  margin-right: 24px;
+}
+
+.reportTypeImage {
+  height: 110px;
+}
+
+.reportImageBackdrop {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 200px;
+  width: 100%;
+  background-color: $grayLightest;
+}
+
+.reportFullDescription {
+  padding: 24px;
+  white-space: pre-wrap;
+}
+
+.reportConfigurationParent {
+  flex-grow: 1;
+  flex-shrink: 1;
+  padding-left: 24px;
+  padding-right: 24px;
+  padding-top: 12px;
+  overflow-y: auto;
+  -webkit-overflow-scrolling: touch;
+}
+
+.reportBackdrop {
+  flex-grow: 1;
+  flex-shrink: 1;
+  background-color: $grayLightest;
+  padding: 24px;
+  overflow-y: auto;
+}
+
+.reportModalWrapper {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
